feat: Add violin plots for peak turning speed
This commit adds violin plots for peak turning speed in the main.asv and main.m files. The plots are generated using the `violin` function and display the turning speed for different task types. The y-axis limits are set to start at 0 and go up to 1000. The plots are displayed in a 1x2 subplot configuration.

The commit also includes the addition of the `Violin_Processing` function to process the data for the violin plots. The function takes in the turning speed data and additional parameters for customizing the plot appearance.

This commit builds upon previous commits that updated the calculations to remove unnecessary NaN values and added a new dataset. It also includes the addition of an excel function to write the data to an excel file, producing the TUG_Data_CSV.

Recent user commits include updates to the README.md file, timing output excel, and the addition of an updated box plot. There were also commits related to adding more DT and ST files, removing the mean calculation, and adding a calculate function and output to excel. Violin Processing added to plot the calc data on violin plots with scatter points